Change setup.py and change methods gift
