+++ .gitignore
@@ -3,3 +3,4 @@ __pycache__
 *.swp
 *.egg-info
 /build
+.venv
