+++ advent_of_code/day1/app.py
