+++ advent_of_code/day1/input.txt
