+++ advent_of_code/day1/santa.py
