+++ setup.py
@@ -4,4 +4,4 @@ from distutils.core import setup
 setup(name='advent_of_code',
       version='0.1',
       description='Advent of code',
-      packages=['day1', 'day2'])
+      packages=['advent_of_code'])
